mongo-api/requirements.txt
certifi==2019.3.9
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pymongo==3.7.2
python-dateutil==2.8.0
six==1.12.0
Werkzeug==0.15.1
